rename ControlBar to topbar
